feat: use default directive definitions if outside amplify project
